multiagent-llm-system/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiagent-llm-system"
version = "0.1.0"
description = "Multiagent LLM system using FastAPI, LangChain, and various LLM providers"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.3.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.24.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
    "prometheus-client>=0.17.0",
    "langchain>=0.0.300",
    "langchain_openai>=0.0.2",
    "langchain_community>=0.0.9",
    "langfuse>=2.0.0",
    "boto3>=1.28.0",
    "jina>=3.15.0",
    "pinecone-client>=2.2.0",
    "weaviate-client>=3.24.0",
    "chromadb>=0.4.0",
    "llama-index>=0.8.0",
    "openai>=1.3.0",
    "websockets>=11.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
